tests fixed as baseline for refactoring
