+++ tellurium/tests/test_kisao.py
@@ -1,5 +1,10 @@
 """
-Testing tephrasedml.
+Testing the support of KISAO terms for SED-ML simulations.
+
+    test_phrasedml.py : phrasedml based tests.
+    test_kisao.py : SED-ML kisao support
+    test_omex.py : SED-ML tests based on Combine Archives
+    test_tesedml.py : tests for the `tesedml.py` module
 """
 from __future__ import absolute_import, print_function
 
@@ -17,13 +22,11 @@ except ImportError:
     import libsedml
 import phrasedml
 
-from tellurium.sedml.tephrasedml import experiment
-import tellurium.sedml.tephrasedml as tephrasedml
-from tellurium.sedml.tesedml import SEDMLCodeFactory
+# from tellurium.sedml.tephrasedml import experiment
+# import tellurium.sedml.tephrasedml as tephrasedml
 
-
-class tePhrasedMLTestCase(unittest.TestCase):
-
+@unittest.skip
+class KisaoSedmlTestCase(unittest.TestCase):
     def tearDown(self):
         matplotlib.pyplot.switch_backend(self.backend)
 
@@ -46,7 +49,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
           task1 = run sim1 on model1
           plot "Figure 1" time vs S1, S2
         '''
-        self.tep = tephrasedml.experiment(self.antimony, self.phrasedml)
+        # self.tep = tephrasedml.experiment(self.antimony, self.phrasedml)
 
         self.a1 = """
         model m1()
@@ -66,100 +69,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
     def tearDown(self):
         self.tep = None
 
-    def test_execute(self):
-        """Test execute."""
-        inline_omex = '\n'.join([self.antimony, self.phrasedml])
-        te.executeInlineOmex(inline_omex)
 
-    def test_exportAsCombine(self):
-        """ Test exportAsCombine. """
-        inline_omex = '\n'.join([self.antimony, self.phrasedml])
-        tmpdir = tempfile.mkdtemp()
-        te.exportInlineOmex(inline_omex, os.path.join(tmpdir, 'archive.omex'))
-        shutil.rmtree(tmpdir)
-
-    def test_1Model1PhrasedML(self):
-        """ Minimal example which should work. """
-        antimony_str = """
-        model test
-            J0: S1 -> S2; k1*S1;
-            S1 = 10.0; S2=0.0;
-            k1 = 0.1;
-        end
-        """
-        phrasedml_str = """
-            model0 = model "test"
-            sim0 = simulate uniform(0, 10, 100)
-            task0 = run sim0 on model0
-            plot task0.time vs task0.S1
-        """
-        inline_omex = '\n'.join([antimony_str, phrasedml_str])
-        te.executeInlineOmex(inline_omex)
-
-    def test_1Model2PhrasedML(self):
-        """ Test multiple models and multiple phrasedml files. """
-        p1 = """
-            model1 = model "m1"
-            sim1 = simulate uniform(0, 6, 100)
-            task1 = run sim1 on model1
-            plot task1.time vs task1.S1, task1.S2
-        """
-
-        p2 = """
-            model1 = model "m1"
-            model2 = model model1 with S1=S2+20
-            sim1 = simulate uniform(0, 6, 100)
-            task1 = run sim1 on model2
-            plot task1.time vs task1.S1, task1.S2
-        """
-        inline_omex = '\n'.join([self.a1, p1])
-        te.executeInlineOmex(inline_omex)
-
-        inline_omex = '\n'.join([self.a1, p2])
-        te.executeInlineOmex(inline_omex)
-
-        inline_omex = '\n'.join([self.a1, p1, p2])
-        te.executeInlineOmex(inline_omex)
-
-    def test_2Model1PhrasedML(self):
-        """ Test multiple models and multiple phrasedml files. """
-        p1 = """
-            model1 = model "m1"
-            model2 = model "m2"
-            model3 = model model1 with S1=S2+20
-            sim1 = simulate uniform(0, 6, 100)
-            task1 = run sim1 on model1
-            task2 = run sim1 on model2
-            plot "Timecourse test1" task1.time vs task1.S1, task1.S2
-            plot "Timecourse test2" task2.time vs task2.X1, task2.X2
-        """
-        inline_omex = '\n'.join([self.a1, self.a2, p1])
-        te.executeInlineOmex(inline_omex)
-
-    def test_2Model2PhrasedML(self):
-        """ Test multiple models and multiple phrasedml files. """
-        p1 = """
-            model1 = model "m1"
-            model2 = model "m2"
-            sim1 = simulate uniform(0, 6, 100)
-            task1 = run sim1 on model1
-            task2 = run sim1 on model2
-            plot task1.time vs task1.S1, task1.S2, task2.time vs task2.X1, task2.X2
-        """
-        p2 = """
-            model1 = model "m1"
-            model2 = model "m2"
-            sim1 = simulate uniform(0, 20, 20)
-            task1 = run sim1 on model1
-            task2 = run sim1 on model2
-            plot task1.time vs task1.S1, task1.S2, task2.time vs task2.X1, task2.X2
-        """
-        inline_omex = '\n'.join([self.a1, self.a2, p1, p2])
-        te.executeInlineOmex(inline_omex)
-
-    ###################################################################################
-    # Testing Kisao Terms
-    ###################################################################################
     def checkKisaoIntegrator(self, exp, kisao, name):
         """ Helper function for checking kisao integrator. """
         p = exp.phrasedmlList[0]
@@ -180,6 +90,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         else:
             self.assertTrue(".setIntegrator('{}')".format(name) in pystr)
 
+
     def test_kisao_cvode_1(self):
         p = """
             model0 = model "m1"
@@ -191,6 +102,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000019', 'cvode')
         exp.execute()
+
 
     def test_kisao_cvode_2(self):
         p = """
@@ -204,6 +116,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoIntegrator(exp, 'KISAO:0000019', 'cvode')
         exp.execute()
 
+
     def test_kisao_cvode_3(self):
         """ Default of uniform is cvode. """
         p = """
@@ -215,6 +128,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000019', 'cvode')
         exp.execute()
+
 
     def test_kisao_cvode_4(self):
         """ Default of onestep is cvode. """
@@ -228,6 +142,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoIntegrator(exp, 'KISAO:0000019', 'cvode')
         exp.execute()
 
+
     def test_kisao_gillespie_1(self):
         p = """
             model0 = model "m1"
@@ -239,6 +154,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000241', 'gillespie')
         exp.execute()
+
 
     def test_kisao_gillespie_2(self):
         p = """
@@ -252,6 +168,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoIntegrator(exp, 'KISAO:0000241', 'gillespie')
         exp.execute()
 
+
     def test_kisao_gillespie_3(self):
         """ Default of uniform_stochastic is gillespie."""
         p = """
@@ -263,6 +180,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000241', 'gillespie')
         exp.execute()
+
 
     def test_kisao_rk4_1(self):
         p = """
@@ -276,6 +194,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoIntegrator(exp, 'KISAO:0000032', 'rk4')
         exp.execute()
 
+
     def test_kisao_rk4_2(self):
         p = """
             model0 = model "m1"
@@ -287,6 +206,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000032', 'rk4')
         exp.execute()
+
 
     # TODO: write tests
     def test_kisao_bdf(self):
@@ -303,6 +223,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         pycode = exp._toPython(p)
         print(pycode)
         self.assertTrue("integrator.setValue('stiff', True)" in pycode)
+
 
     def test_kisao_adams(self):
         p = """
@@ -332,6 +253,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoIntegrator(exp, 'KISAO:0000435', 'rk45')
         exp.execute()
+
 
     @pytest.mark.skip(reason="bug in roadrunner")
     def test_kisao_rk45_2(self):
@@ -387,6 +309,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
                 # numerical parameter
                 self.assertTrue(".integrator.setValue('{}', {})".format(name, value) in pystr)
 
+
     def test_kisao_relative_tolerance_1(self):
         p = """
             model0 = model "m1"
@@ -398,6 +321,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000209', 'relative_tolerance', 1E-8)
         exp.execute()
+
 
     def test_kisao_relative_tolerance_2(self):
         p = """
@@ -411,6 +335,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000209', 'relative_tolerance', 1E-8)
         exp.execute()
 
+
     def test_kisao_absolute_tolerance_1(self):
         p = """
             model0 = model "m1"
@@ -422,6 +347,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000211', 'absolute_tolerance', 1E-8)
         exp.execute()
+
 
     def test_kisao_absolute_tolerance_2(self):
         p = """
@@ -435,6 +361,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000211', 'absolute_tolerance', 1E-8)
         exp.execute()
 
+
     def test_kisao_maximum_bdf_order_1(self):
         p = """
             model0 = model "m1"
@@ -446,6 +373,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000220', 'maximum_bdf_order', 4)
         exp.execute()
+
 
     def test_kisao_maximum_bdf_order_2(self):
         p = """
@@ -459,6 +387,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000220', 'maximum_bdf_order', 4)
         exp.execute()
 
+
     def test_kisao_maximum_adams_order_1(self):
         p = """
             model0 = model "m1"
@@ -470,6 +399,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000219', 'maximum_adams_order', 5)
         exp.execute()
+
 
     def test_kisao_maximum_adams_order_2(self):
         p = """
@@ -483,6 +413,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000219', 'maximum_adams_order', 5)
         exp.execute()
 
+
     def test_kisao_maximum_num_steps_1(self):
         p = """
             model0 = model "m1"
@@ -494,6 +425,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000415', 'maximum_num_steps', 10000)
         exp.execute()
+
 
     def test_kisao_maximum_num_steps_2(self):
         p = """
@@ -520,6 +452,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000467', 'maximum_time_step', 1.0)
         exp.execute()
 
+
     def test_kisao_maximum_time_step_2(self):
         p = """
             model0 = model "m1"
@@ -531,6 +464,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000467', 'maximum_time_step', 1.0)
         exp.execute()
+
 
     @pytest.mark.skip(reason="bug in roadrunner")
     def test_kisao_minimum_time_step_1(self):
@@ -545,6 +479,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000485', 'minimum_time_step', 1E-6)
         exp.execute()
 
+
     @pytest.mark.skip(reason="bug in roadrunner")
     def test_kisao_minimum_time_step_2(self):
         p = """
@@ -558,6 +493,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000485', 'minimum_time_step', 1E-6)
         exp.execute()
 
+
     def test_kisao_initial_time_step_1(self):
         p = """
             model0 = model "m1"
@@ -569,6 +505,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000332', 'initial_time_step', 0.01)
         exp.execute()
+
 
     def test_kisao_initial_time_step_2(self):
         p = """
@@ -582,6 +519,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000332', 'initial_time_step', 0.01)
         exp.execute()
 
+
     def test_kisao_variable_step_size_1(self):
         p = """
             model0 = model "m1"
@@ -593,6 +531,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000107', 'variable_step_size', True)
         exp.execute()
+
 
     def test_kisao_variable_step_size_2(self):
         p = """
@@ -619,6 +558,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000486', 'maximum_iterations', 10)
         exp.execute()
 
+
     def test_kisao_maximum_iterations_2(self):
         p = """
             model0 = model "m1"
@@ -631,6 +571,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000486', 'maximum_iterations', 10)
         exp.execute()
 
+
     def test_kisao_minimum_damping_1(self):
         p = """
             model0 = model "m1"
@@ -642,6 +583,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000487', 'minimum_damping', 1.0)
         exp.execute()
+
 
     def test_kisao_minimum_damping_2(self):
         p = """
@@ -656,6 +598,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         print(exp._toPython(p))
         exp.execute()
 
+
     def test_kisao_seed_1(self):
         p = """
             model0 = model "m1"
@@ -668,6 +611,7 @@ class tePhrasedMLTestCase(unittest.TestCase):
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000488', 'seed', 1234)
         exp.execute()
 
+
     def test_kisao_seed_2(self):
         p = """
             model0 = model "m1"
@@ -679,7 +623,3 @@ class tePhrasedMLTestCase(unittest.TestCase):
         exp = experiment(self.a1, p)
         self.checkKisaoAlgorithmParameter(exp, 'KISAO:0000488', 'seed', 1234)
         exp.execute()
-
-
-if __name__ == '__main__':
-    unittest.main()
