+++ tellurium/tests/test_omex.py
@@ -1,0 +1,8 @@
+"""
+Test that Combine archives can be executed successfully.
+
+    test_phrasedml.py : phrasedml based tests.
+    test_kisao.py : SED-ML kisao support
+    test_omex.py : SED-ML tests based on Combine Archives
+    test_tesedml.py : tests for the `tesedml.py` module
+"""
